Removed ViewModel factory from interactor
