+++ Game of Life/Game of Life/modules/showHomeScreen/HomeContent.swift
@@ -1,0 +1,7 @@
+import Foundation
+
+struct HomeContent {
+  let header = "Game of Life"
+  let startPrompt = "Start"
+  let settingsPrompt = "Settings"
+}
