+++ Game of Life/Game of Life/modules/showHomeScreen/HomeViewModel.swift
@@ -1,14 +1,5 @@
 import Foundation
 
-protocol HomeViewModelFactory {
-  func create() -> HomeViewModel
-}
-
 struct HomeViewModel {
-  struct ButtonData: Hashable {
-    let title: String
-  }
-  
-  let buttons: [ButtonData]
-  let particleEffect: URL?
+  let content = HomeContent()
 }
