+++ Game of Life/Game of Life/modules/showSettings/SettingsContent.swift
@@ -1,0 +1,7 @@
+import Foundation
+
+struct SettingsContent {
+  let header = "Settings"
+  let changeGridSizePrompt = "Grid size"
+  let changeSpeedPrompt = "Speed"
+}
