+++ Game of Life/Game of Life/modules/showSettings/SettingsScreenView.swift
@@ -11,18 +11,23 @@ struct SettingsScreenView: View {
   
   var body: some View {
     ZStack(alignment: .top) {
-      PlayerView(url: interactor.viewModel.particleEffect)
-        .frame(maxWidth: .infinity, maxHeight: .infinity)
+      PlayerView(
+        url: Bundle.main.url(
+          forResource: "particle-effect",
+          withExtension: ".mp4"
+        )
+      )
+      .frame(maxWidth: .infinity, maxHeight: .infinity)
       
       ZStack(alignment: .leading) {
-        Button(action: onBackButtonTapped)  {
+        Button(action: interactor.closeSettings)  {
           Image("back-arrow")
             .resizable()
             .aspectRatio(contentMode: .fit)
             .tint(AppAppearance.Colors.color_313031)
             .frame(width: 22.5, height: 22.5)
         }
-        Text("Settings")
+        Text(interactor.viewModel.content.header)
           .font(AppAppearance.Fonts.medium_18)
           .frame(maxWidth: .infinity)
       }
@@ -31,13 +36,13 @@ struct SettingsScreenView: View {
       
       VStack(alignment: .leading, spacing: AppAppearance.Spacing.large) {
         slider(
-          title: "Grid size",
+          title: interactor.viewModel.content.changeGridSizePrompt,
           value: $interactor.viewModel.grid.size,
           minValue: interactor.viewModel.grid.minSize,
           maxValue: interactor.viewModel.grid.maxSize
         )
         slider(
-          title: "Speed",
+          title: interactor.viewModel.content.changeSpeedPrompt,
           value: $interactor.viewModel.speed.speed,
           minValue: interactor.viewModel.speed.minSpeed,
           maxValue: interactor.viewModel.speed.maxSpeed
@@ -48,15 +53,13 @@ struct SettingsScreenView: View {
     .frame(maxWidth: .infinity, maxHeight: .infinity)
     .background(AppAppearance.Colors.color_ffffff)
   }
-  
-  // MARK: - Private methods
-  
-  private func onBackButtonTapped() {
-    interactor.closeSettings()
-  }
-  
+}
+
+// MARK: - SettingsScreenView+UI
+
+private extension SettingsScreenView {
   @ViewBuilder
-  private func slider(
+  func slider(
     title: String,
     value: Binding<Int>,
     minValue: Int,
@@ -67,7 +70,7 @@ struct SettingsScreenView: View {
         .font(AppAppearance.Fonts.regular_16)
         .offset(y: -7)
       Spacer()
-      Slider(
+      SliderView(
         value: value,
         min: minValue,
         max: maxValue
@@ -75,74 +78,6 @@ struct SettingsScreenView: View {
       .frame(width: 250, height: 46)
     }
     .frame(width: 340)
-  }
-}
-
-// MARK: - SettingsScreenView+Slider
-
-private extension SettingsScreenView {
-  struct Slider: View {
-    // MARK: - Datasource properties
-    
-    @Binding
-    var value: Int
-    @State
-    private var initialDragValue: Int?
-    let min: Int
-    let max: Int
-    var thumbSize: CGFloat = 30
-    
-    // MARK: - Body
-    
-    var body: some View {
-      GeometryReader { proxy in
-        ZStack(alignment: .top) {
-          AppAppearance.Colors.color_313031
-            .frame(height: 3)
-            .offset(y: thumbSize / 2)
-          
-          VStack(spacing: AppAppearance.Spacing.extraSmall) {
-            AppAppearance.Colors.color_313031
-              .frame(width: 12.5, height: thumbSize)
-            Text("\(value)")
-              .font(AppAppearance.Fonts.light_13)
-          }
-          .offset(x: -proxy.size.width/2)
-          .offset(x: getDragProgress(with: value)*proxy.size.width)
-        }
-        .gesture(
-          DragGesture()
-            .onChanged({
-              onThumbDrag(
-                offset: $0.translation.width,
-                viewWidth: proxy.size.width
-              )
-            })
-            .onEnded({ _ in
-              initialDragValue = nil
-            })
-        )
-      }
-    }
-    
-    // MARK: - Private methods
-    
-    private func onThumbDrag(offset: Double, viewWidth: Double) {
-      if initialDragValue == nil {
-        initialDragValue = value
-      }
-      
-      let dragProgress = abs(offset) / viewWidth
-      let dragMultiplier = offset < 0 ? -1 : 1
-      let dragDelta = dragProgress * Double(max - min)
-      let newValue = (initialDragValue ?? 0) + dragMultiplier*Int(dragDelta)
-      
-      value = Swift.min(max, Swift.max(min, newValue))
-    }
-    
-    private func getDragProgress(with value: Int) -> Double {
-      return Double(value - min) / Double(max-min)
-    }
   }
 }
