+++ Game of Life/Game of Life/modules/showSettings/SettingsViewModel.swift
@@ -1,9 +1,5 @@
 import Foundation
 
-protocol SettingsViewModelFactory {
-  func create(with settings: AppState.Settings) -> SettingsViewModel
-}
-
 struct SettingsViewModel {
   struct GridMetadata {
     var size: Int
@@ -16,8 +12,8 @@ struct SettingsViewModel {
     let minSpeed: Int
   }
   
+  let content = SettingsContent()
   var grid: GridMetadata
   var speed: SpeedMetadata
-  let particleEffect: URL?
 }
